Use sqlite to store message IDs
