+++ .gitignore
@@ -1,3 +1,5 @@
+db.sqlite
+
 messages/
 logs/
 config/
