+++ hll_server_status/cli.py
@@ -9,12 +9,11 @@ from loguru import logger
 from hll_server_status import constants
 from hll_server_status.io import (
     load_config,
-    load_message_ids,
     queue_webhook_update,
-    save_message_ids_to_disk,
     send_queued_webhook_update,
 )
-from hll_server_status.models import AppStore
+from hll_server_status.models import enter_session, get_set_wh_row
+from hll_server_status.types import AppStore
 
 # Disable logging so discord_webhook doesn't log for us
 logging.config.dictConfig(
@@ -52,9 +51,8 @@ async def main():
         app_store = AppStore(
             server_identifier=config_file_path.stem,
             logger=_logger,
-            last_saved_message_ids=None,
         )
-        await load_message_ids(app_store)
+
         config_files.append((app_store, config_file_path))
 
     if not config_files:
@@ -65,75 +63,59 @@ async def main():
     toml_section_keys = (
         "header",
         "gamestate",
-        "map_rotation_color",
-        "map_rotation_embed",
+        "map_rotation",
         "player_stats",
     )
-    table_name = constants.MESSAGE_ID_FORMAT["table_name"]
-
     # Use a 0 size buffer so we never queue another attempt until the previous one has been
     # received since these are all snap shots and producing faster than we can consume is
     # negative value
-    try:
-        send_channel, receive_channel = trio.open_memory_channel(0)
-        async with trio.open_nursery() as nursery:
-            async with send_channel, receive_channel:
-                for app_store, config_file_path in config_files:
-                    default_logger.info(
-                        f"Starting {config_file_path} check log files for further output"
+    send_channel, receive_channel = trio.open_memory_channel(0)
+    async with trio.open_nursery() as nursery:
+        async with send_channel, receive_channel:
+            for app_store, config_file_path in config_files:
+                default_logger.info(
+                    f"Starting {config_file_path} check log files for further output"
+                )
+                print(f"Starting {config_file_path} check log files for further output")
+
+                app_store.logger.info(f"Reading config file for {config_file_path}")
+                try:
+                    config = load_config(app_store, config_file_path)
+                except (KeyError, ValueError) as e:
+                    app_store.logger.error(
+                        f"{e} while loading config from {config_file_path}"
                     )
-                    print(
-                        f"Starting {config_file_path} check log files for further output"
+                    continue
+
+                logger.info(f"Testing a connection to {config.api.base_server_url}")
+                try:
+                    async with httpx.AsyncClient() as client:
+                        await client.get(str(config.api.base_server_url))
+                except httpx.ConnectError as e:
+                    app_store.logger.error(
+                        f"Unable to connect to {config.api.base_server_url} for {config_file_path}"
                     )
 
-                    app_store.logger.info(f"Reading config file for {config_file_path}")
-                    try:
-                        config = load_config(app_store, config_file_path)
-                    except (KeyError, ValueError) as e:
-                        app_store.logger.error(
-                            f"{e} while loading config from {config_file_path}"
-                        )
-                        continue
+                for section_key in toml_section_keys:
+                    job_key = f"{app_store.server_identifier}:{section_key}"
 
-                    logger.info(f"Testing a connection to {config.api.base_server_url}")
-                    try:
-                        async with httpx.AsyncClient() as client:
-                            await client.get(str(config.api.base_server_url))
-                    except httpx.ConnectError as e:
-                        app_store.logger.error(
-                            f"Unable to connect to {config.api.base_server_url} for {config_file_path}"
-                        )
+                    # Create a unique queue for each section in each config file so they can all update
+                    # independently of each other
+                    send_channel_clone = send_channel.clone()
+                    receive_channel_clone = receive_channel.clone()
 
-                    for section_key in toml_section_keys:
-                        job_key = f"{app_store.server_identifier}:{section_key}"
-
-                        # Create a unique queue for each section in each config file so they can all update
-                        # independently of each other
-                        send_channel_clone = send_channel.clone()
-                        receive_channel_clone = receive_channel.clone()
-
-                        nursery.start_soon(
-                            queue_webhook_update,
-                            send_channel_clone,
-                            job_key,
-                            config,
-                            config_file_path,
-                            app_store,
-                            table_name,
-                            section_key,
-                        )
-                        nursery.start_soon(
-                            send_queued_webhook_update, receive_channel_clone, job_key
-                        )
-    except Exception as e:
-        raise e
-    finally:
-        # Make one final attempt to save all of our message IDs to disk
-        # otherwise unlucky KeyboardInterrupt timings or other exceptions
-        # can cause them to not be written after a message has been created
-        # and leave them orphaned in Discord, which users hate!
-        for app_store, _ in config_files:
-            await save_message_ids_to_disk(app_store, None)
+                    nursery.start_soon(
+                        queue_webhook_update,
+                        send_channel_clone,
+                        job_key,
+                        config,
+                        config_file_path,
+                        app_store,
+                        section_key,
+                    )
+                    nursery.start_soon(
+                        send_queued_webhook_update, receive_channel_clone, job_key
+                    )
 
 
 if __name__ == "__main__":
